// src/cyfs-stack-test-typescript/cyfs_node/cyfs_node.d.ts
export class Tx extends NamedObject<MetaTxDescContentType, TxBodyContent> {
    static create(nonce: JSBI, caller: TxCaller, gas_coin_id: number, gas_price: number, max_fee: number, condition: Option<TxCondition>, body: MetaTxBody, data: Uint8Array): Tx;
    ext(): TxExt;
}
// ...
    interface Uint8Array {
        offset(offset: number): Uint8Array;
        offsetUint16Array(offset: number): Uint16Array;
        offsetUint32Array(offset: number): Uint32Array;
        offsetView(offset: number): DataView;
        toHex(): string;
        fromHex(hex: string): BuckyResult<Uint8Array>;
    }
// ...
export enum RawEncodePurpose {
        Serialize = 0,
        Hash = 1
}
export interface RawEncode {
        raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf?: (purpose?: RawEncodePurpose) => BuckyResult<Uint8Array>;
        raw_hash_encode?: () => BuckyResult<HashValue>;
}
// ...
export interface Compareable<T> {
        hashCode(): symbol;
        equals(other: T): boolean;
}
// ...
export type BuckyResult<T> = Result<T, BuckyError>;
// ...
export class HashValue implements RawEncode {
    m_buf: Uint8Array;
    constructor(buf: Uint8Array);
    as_slice(): Uint8Array;
    length(): number;
    static default(): HashValue;
    static copy_from_slice(buf: Uint8Array): HashValue;
    static hash_data(data: Uint8Array): HashValue;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    to_base_58(): string;
    to_hex_string(): string;
    static from_hex_string(s: string): BuckyResult<HashValue>;
}
// ...
export class AppGroup extends NamedObject<AppGroupDescContent, AppGroupBodyContent> {
    static create(build?: (builder: AppGroupBuilder) => void): AppGroup;
    appgroup_id(): AppGroupId;
    connect_info(): AppGroupBodyContent;
}
// ...
export class Device extends NamedObject<DeviceDescContent, DeviceBodyContent> {
    static create(owner: Option<ObjectId>, unique_id: UniqueId, endpoints: Endpoint[], sn_list: DeviceId[], passive_sn_list: DeviceId[], public_key: PublicKey, area: Area, category: DeviceCategory, build?: (builder: DeviceBuilder) => void): Device;
    device_id(): DeviceId;
    connect_info(): DeviceBodyContent;
    name(): string | undefined;
    set_name(name?: string): void;
    category(): BuckyResult<DeviceCategory>;
}
// ...
export class Dir extends NamedObject<DirDescContent, DirBodyContent> {
    static create(owner: ObjectId, attributes: Attributes, obj_list: NDNObjectInfo, body: {
        chunk_id?: ChunkId;
        obj_list?: BuckyHashMap<ObjectId, BuckyBuffer>;
    }, build?: (builder: DirBuilder) => void): Dir;
    dir_id(): DirId;
    get_data_from_body(id: ObjectId): Option<Uint8Array>;
    check_and_fix_desc_limit(): BuckyResult<void>;
}
// ...
export class File extends NamedObject<FileDescContent, FileBodyContent> {
    static create(owner: ObjectId, len: JSBI, hash: HashValue, chunk_list: ChunkList, build?: (builder: FileBuilder) => void): File;
    file_id(): FileId;
}
// ...
export class NamedObjectId<DC extends DescContent, BC extends BodyContent> implements RawEncode, Compareable<NamedObjectId<DC, BC>> {
        constructor(obj_type: number, object_id: ObjectId);
        get object_id(): ObjectId;
        hashCode(): symbol;
        equals<ODC extends DescContent, OBC extends BodyContent>(other: NamedObjectId<ODC, OBC>): boolean;
        into(): ObjectId;
        gen(object_id: ObjectId): void;
        toString(): string;
        toJSON(): string;
        to_string(): string;
        to_base_58(): string;
        raw_measure(ctx?: any): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
// ...
export class NamedObjectIdDecoder<DC extends DescContent, BC extends BodyContent> implements RawDecode<NamedObjectId<DC, BC>> {
        readonly m_obj_type: number;
        constructor(obj_type: number);
        raw_decode(buf: Uint8Array, ctx?: any): BuckyResult<[NamedObjectId<DC, BC>, Uint8Array]>;
}
// ...
export abstract class DescTypeInfo {
        get_sub_obj_type(): number;
        set_sub_obj_type(v: number): void;
        abstract obj_type(): number;
        abstract sub_desc_type(): SubDescType;
}
export abstract class DescContent implements RawEncode {
        abstract type_info(): DescTypeInfo;
        codec_info(): ContentCodecInfo;
        abstract raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        abstract raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
}
export abstract class DescContentDecoder<T extends DescContent> implements RawDecode<T> {
        abstract type_info(): DescTypeInfo;
        abstract raw_decode(buf: Uint8Array, ctx: ContentRawDecodeContext): BuckyResult<[T, Uint8Array]>;
}
// ...
export abstract class BodyContent implements RawEncode {
        codec_info(): ContentCodecInfo;
        abstract raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        abstract raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
}
export abstract class BodyContentDecoder<T extends BodyContent> implements RawDecode<T> {
        abstract raw_decode(buf: Uint8Array, ctx: ContentRawDecodeContext): BuckyResult<[T, Uint8Array]>;
}
// ...
export class NamedObjectDescDecoder<T extends DescContent> implements RawDecode<NamedObjectDesc<T>> {
        constructor(desc_content_decoder: DescContentDecoder<T>);
        trace_id(): number;
        raw_decode(buf: Uint8Array, ctx?: NamedObjectContext): BuckyResult<[NamedObjectDesc<T>, Uint8Array]>;
}
export class NamedObject<DC extends DescContent, BC extends BodyContent> implements RawEncode {
        constructor(desc: NamedObjectDesc<DC>, body: Option<ObjectMutBody<DC, BC>>, signs: ObjectSigns, nonce: Option<JSBI>);
        obj_type(): number;
        obj_type_code(): number;
        calculate_id(): ObjectId;
        protected obj_flags(): number;
        to_vec(): BuckyResult<Uint8Array>;
        to_hex(): BuckyResult<string>;
        toString(): string;
        to_string(): string;
        toJSON(): string;
        desc(): NamedObjectDesc<DC>;
        body(): Option<ObjectMutBody<DC, BC>>;
        set_body(body: Option<ObjectMutBody<DC, BC>>): void;
        body_expect(): ObjectMutBody<DC, BC>;
        signs(): ObjectSigns;
        nonce(): Option<JSBI>;
        raw_measure(_ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: NamedObjectContext, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf(purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        raw_hash_encode(): BuckyResult<HashValue>;
}
export class NamedObjectDecoder<DC extends DescContent, BC extends BodyContent, O extends NamedObject<DC, BC>> implements RawDecode<NamedObject<DC, BC>> {
        constructor(desc_content_decoer: DescContentDecoder<DC>, body_content_decoder: BodyContentDecoder<BC>, obj_builder: new (...constructorArgs: any[]) => O);
        raw_decode(buf: Uint8Array): BuckyResult<[O, Uint8Array]>;
        from_raw(buf: Uint8Array): BuckyResult<O>;
        from_hex(hex: string): BuckyResult<O>;
}
// ...
export enum ObjectCategory {
    Standard = "standard",
    Core = "core",
    DecApp = "dec_app"
}
// ...
export interface ObjectIdInfoMatcher {
    match<T>(p: ObjectIdInfoPartten<T>): T;
}
// ...
export type ObjectIdInfo = (StandardObjectIdInfo | CoreObjectIdInfo | DecAppObjectIdInfo) & ObjectIdInfoMatcher;
export class ObjectId implements RawEncode, Compareable<ObjectId> {
    m_buf: Uint8Array;
    m_base58?: string;
    constructor(buf: Uint8Array);
    get object_id(): void;
    static default(): ObjectId;
    static copy_from_slice(buf: Uint8Array): ObjectId;
    obj_type_code(): ObjectTypeCode;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    as_slice(): Uint8Array;
    clone(): ObjectId;
    length(): number;
    toString(): string;
    toJSON(): string;
    to_string(): string;
    to_base_58(): string;
    hashCode(): symbol;
    equals(other: ObjectId): boolean;
    eq(other: ObjectId): boolean;
    static from_str(s: string): BuckyResult<ObjectId>;
    static from_base_58(s: string): BuckyResult<ObjectId>;
    to_hash_value(): HashValue;
    object_category(): ObjectCategory;
    is_stand_object(): boolean;
    is_core_object(): boolean;
    is_dec_app_object(): boolean;
    info(): ObjectIdInfo;
}
// ...
export enum ObjectTypeCode {
    Device = 1,
    People = 2,
    SimpleGroup = 3,
    Org = 4,
    AppGroup = 5,
    UnionAccount = 6,
    Chunk = 7,
    File = 8,
    Dir = 9,
    Diff = 10,
    ProofOfService = 11,
    Tx = 12,
    Action = 13,
    ObjectMap = 14,
    Contract = 15,
    Custom = 16
}
// ...
export class Org extends NamedObject<OrgDescContent, OrgBodyContent> {
    static create(members: OrgMember[], directors: Director[], total_equity: string | number, build?: (builder: OrgBuilder) => void): Org;
    org_id(): OrgId;
    content_info(): OrgBodyContent;
}
// ...
export class People extends NamedObject<PeopleDescContent, PeopleBodyContent> {
    static create(owner: Option<ObjectId>, ood_list: DeviceId[], public_key: PublicKey, area: Option<Area>, name?: string, icon?: FileId, build?: (builder: PeopleBuilder) => void): People;
    people_id(): PeopleId;
    name(): string | undefined;
    set_name(name?: string): void;
    ood_work_mode(): OODWorkMode;
    set_ood_work_mode(mode: OODWorkMode): void;
    icon(): FileId | undefined;
    set_icon(icon?: FileId): void;
}
// ...
export class SimpleGroup extends NamedObject<SimpleGroupDescContent, SimpleGroupBodyContent> {
    static create(threshold: number, owners: PublicKey[], members: ObjectId[], ood_work_mode: OODWorkMode, ood_list: DeviceId[], area: Area, build?: (builder: SimpleGroupBuilder) => void): SimpleGroup;
    ood_work_mode(): OODWorkMode;
    set_ood_work_mode(mode: OODWorkMode): void;
    simple_group_id(): SimpleGroupId;
    connect_info(): SimpleGroupBodyContent;
}
export class SimpleGroupDecoder extends NamedObjectDecoder<SimpleGroupDescContent, SimpleGroupBodyContent, SimpleGroup> {
    constructor();
}

export class UnionAccountDescTypeInfo extends DescTypeInfo {
    obj_type(): number;
    sub_desc_type(): SubDescType;
}
export class UnionAccountDescContent extends DescContent {
    left: ObjectId;
    right: ObjectId;
    service_type: number;
    constructor(left: ObjectId, right: ObjectId, service_type: number);
    type_info(): DescTypeInfo;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class UnionAccountDescContentDecoder extends DescContentDecoder<UnionAccountDescContent> {
    type_info(): DescTypeInfo;
    raw_decode(buf: Uint8Array): BuckyResult<[UnionAccountDescContent, Uint8Array]>;
}
export class UnionAccountBodyContent extends BodyContent {
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class UnionAccountBodyContentDecoder extends BodyContentDecoder<UnionAccountBodyContent> {
    raw_decode(buf: Uint8Array): BuckyResult<[UnionAccountBodyContent, Uint8Array]>;
}
export class UnionAccountDesc extends NamedObjectDesc<UnionAccountDescContent> {
}
export class UnionAccountDescDecoder extends NamedObjectDescDecoder<UnionAccountDescContent> {
    constructor();
}
export class UnionAccountBuilder extends NamedObjectBuilder<UnionAccountDescContent, UnionAccountBodyContent> {
}
export class UnionAccountId extends NamedObjectId<UnionAccountDescContent, UnionAccountBodyContent> {
    constructor(id: ObjectId);
    static default(): UnionAccountId;
    static from_base_58(s: string): BuckyResult<UnionAccountId>;
    static try_from_object_id(id: ObjectId): BuckyResult<UnionAccountId>;
}
export class UnionAccountIdDecoder extends NamedObjectIdDecoder<UnionAccountDescContent, UnionAccountBodyContent> {
    constructor();
}
export class UnionAccount extends NamedObject<UnionAccountDescContent, UnionAccountBodyContent> {
    static create(account1: ObjectId, account2: ObjectId, service_type: number, build?: (builder: UnionAccountBuilder) => void): UnionAccount;
    union_account_id(): UnionAccountId;
    connect_info(): UnionAccountBodyContent;
}
export class UnionAccountDecoder extends NamedObjectDecoder<UnionAccountDescContent, UnionAccountBodyContent, UnionAccount> {
    constructor();
}
// ...
export class ProofOfService extends NamedObject<ProofOfServiceDescContent, ProofOfServiceBodyContent> {
    static create(owner: Option<ObjectId>, proof_type: ProofTypeCode, data_0: ProofData, data_1: ProofData): ProofOfService;
    ext(): ProofOfServiceExt;
}